wip: partial commit
comitting this because i'm likely dropping the project, unless
new ideas hit me.

In short, navigation of source code via a generic AST is proving
difficult to conceptualize. Furthermore, the more i explore
navigation of the AST, it seems that what i really want is not an
AST but an editor focused more around templating.

Templating definitely needs some more consideration, but currently
it's not of interest to me. Perhaps in the future i'll follow it
more.